Add files via upload
in the above analysis we have used supervised Machine learning Linear regression algorithm to predict the score of a student on the basis of no. of hours studied.
Data were provided by The spark foundation.we predicted if student studied for 9.25 hours/day, he/she may score around 93 marks. with a mean error of 4%